+++ src/adapters/base.py
