+++ src/adapters/stock.py
